moved getPerformer function to helpers folder
